move Doctrine Rectors to own package
