+++ packages/Doctrine/tests/Rector/AliasToClassRector/AliasToClassRectorTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\Contrib\Doctrine\AliasToClassRector;
+namespace Rector\Doctrine\Tests\Rector\AliasToClassRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
