+++ packages/Doctrine/tests/Rector/AliasToClassRector/Correct/correct.php.inc
