+++ packages/Doctrine/tests/Rector/AliasToClassRector/Wrong/wrong.php.inc
