+++ packages/Doctrine/tests/Rector/AliasToClassRector/config.yml
@@ -1,4 +1,4 @@
 services:
-    Rector\Rector\Contrib\Doctrine\AliasToClassRector:
+    Rector\Doctrine\Rector\AliasToClassRector:
       $aliasesToNamespaces:
         'App': App\Entity
